New board: Fluff M0 Headless
This is the same as Fluff M0, but disables the MSC, CDC and MIDI USB
endpoints. This is useful for devices like keyboards or mice running
CircuitPython, once we are happy with their code and don't want to
change it anymore.

I'm adding this board mostly so that we have a test for this case.